# backend/接口文档.openapi (2).json
{
  "openapi": "3.1.0",
  "info": {
    "title": "默认模块",
    "description": "校园设备管理系统",
    "version": "1.0.0"
  },
...
  "paths": {
    "/user/equipment/page": {
      "get": {
        "summary": "分页条件查询设备",
        "deprecated": false,
        "description": "",
        "operationId": "分页条件查询设备_user_equipment_page_get",
        "tags": [
          "学生端/设备相关"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "页码",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页条数",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "每页条数",
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "设备分类ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备分类ID",
              "title": "Categoryid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "设备状态",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备状态",
              "title": "Status"
            }
          },
          {
            "name": "equipmentName",
            "in": "query",
            "description": "设备名称",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备名称",
              "title": "Equipmentname"
            }
          },
          {
            "name": "equipmentNo",
            "in": "query",
            "description": "设备编号",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备编号",
              "title": "Equipmentno"
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "设备存放位置",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备存放位置",
              "title": "Location"
            }
          },
          {
            "name": "brand",
            "in": "query",
            "description": "设备品牌",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备品牌",
              "title": "Brand"
            }
          },
          {
            "name": "spec",
            "in": "query",
            "description": "设备规格型号",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备规格型号",
              "title": "Spec"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "设备采购开始时间",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备采购开始时间",
              "title": "Starttime"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "设备采购结束时间",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备采购结束时间",
              "title": "Endtime"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "排序字段",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "排序字段",
              "default": "id",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "排序顺序",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "排序顺序",
              "default": "asc",
              "title": "Order"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Page_EquipmentOut__"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/equipment/{equipmentId}": {
      "get": {
        "summary": "根据 Id 查询设备详情",
        "deprecated": false,
        "description": "",
        "operationId": "根据_ID_查询设备详情_user_equipment__equipmentId__get",
        "tags": [
          "学生端/设备相关"
        ],
        "parameters": [
          {
            "name": "equipmentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": 35,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Equipmentid"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_EquipmentOut_"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/repair/equipment/page": {
      "get": {
        "summary": "分页条件查询设备",
        "deprecated": false,
        "description": "",
        "operationId": "分页条件查询设备_repair_equipment_page_get",
        "tags": [
          "维修端/设备相关"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "页码",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页条数",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "每页条数",
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "设备分类ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备分类ID",
              "title": "Categoryid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "设备状态",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备状态",
              "title": "Status"
            }
          },
          {
            "name": "equipmentName",
            "in": "query",
            "description": "设备名称",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备名称",
              "title": "Equipmentname"
            }
          },
          {
            "name": "equipmentNo",
            "in": "query",
            "description": "设备编号",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备编号",
              "title": "Equipmentno"
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "设备存放位置",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备存放位置",
              "title": "Location"
            }
          },
          {
            "name": "brand",
            "in": "query",
            "description": "设备品牌",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备品牌",
              "title": "Brand"
            }
          },
          {
            "name": "spec",
            "in": "query",
            "description": "设备规格型号",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备规格型号",
              "title": "Spec"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "设备采购开始时间",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备采购开始时间",
              "title": "Starttime"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "设备采购结束时间",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备采购结束时间",
              "title": "Endtime"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "排序字段",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "排序字段",
              "default": "id",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "排序顺序",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "排序顺序",
              "default": "asc",
              "title": "Order"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Page_EquipmentOut__"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/repair/equipment/{equipmentId}": {
      "get": {
        "summary": "根据 Id 查询设备详情",
        "deprecated": false,
        "description": "",
        "operationId": "根据_ID_查询设备详情_repair_equipment__equipmentId__get",
        "tags": [
          "维修端/设备相关"
        ],
        "parameters": [
          {
            "name": "equipmentId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Equipmentid"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_EquipmentOut_"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/equipment/page": {
      "get": {
        "summary": "分页条件查询设备",
        "deprecated": false,
        "description": "",
        "operationId": "分页条件查询设备_admin_equipment_page_get",
        "tags": [
          "管理端/设备相关"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "页码",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页条数",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "每页条数",
              "default": 10,
              "title": "Size"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "设备分类ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备分类ID",
              "title": "Categoryid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "设备状态",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备状态",
              "title": "Status"
            }
          },
          {
            "name": "equipmentName",
            "in": "query",
            "description": "设备名称",
            "required": false,
            "example": "测试设备",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备名称",
              "title": "Equipmentname"
            }
          },
          {
            "name": "equipmentNo",
            "in": "query",
            "description": "设备编号",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备编号",
              "title": "Equipmentno"
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "设备存放位置",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备存放位置",
              "title": "Location"
            }
          },
          {
            "name": "brand",
            "in": "query",
            "description": "设备品牌",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备品牌",
              "title": "Brand"
            }
          },
          {
            "name": "spec",
            "in": "query",
            "description": "设备规格型号",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备规格型号",
              "title": "Spec"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "设备采购开始时间",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备采购开始时间",
              "title": "Starttime"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "设备采购结束时间",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "设备采购结束时间",
              "title": "Endtime"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "排序字段",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "排序字段",
              "default": "id",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "排序顺序",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "排序顺序",
              "default": "asc",
              "title": "Order"
            }
          },
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result_Page_EquipmentOut__"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/equipment": {
      "post": {
        "summary": "新增设备",
        "deprecated": false,
        "description": "",
        "operationId": "新增设备_admin_equipment_post",
        "tags": [
          "管理端/设备相关"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "",
            "required": false,
            "example": "",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "Result_Page_EquipmentOut__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务状态码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page_EquipmentOut_"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[Page[EquipmentOut]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "EquipmentCreate": {
        "properties": {
          "equipmentNo": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Equipmentno",
            "description": "设备编号"
          },
          "equipmentName": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Equipmentname",
            "description": "设备名称"
          },
          "categoryId": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoryid",
            "description": "设备分类 ID"
          },
          "spec": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Spec",
            "description": "设备规格型号"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand",
            "description": "设备品牌"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit",
            "description": "计量单位"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "设备存放位置"
          },
          "purchaseDate": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchasedate",
            "description": "采购日期"
          },
          "price": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,8}|(?=[\\d.]{1,11}0*$)\\d{0,8}\\.\\d{0,2}0*$)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price",
            "description": "采购价格"
          },
          "coverImg": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverimg",
            "description": "设备封面图片"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "设备状态",
            "default": "available"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark",
            "description": "备注"
          }
        },
        "type": "object",
        "required": [
          "equipmentNo",
          "equipmentName"
        ],
        "title": "EquipmentCreate",
        "description": "新增设备请求模型"
      },
...
      "Result_EquipmentOut_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "业务状态码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "响应消息"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EquipmentOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "响应数据"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Result[EquipmentOut]"
      },
...
    },
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}